[REFACTOR] 메서드 순서나 이름을 변경하였습니다.
