+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -5,7 +5,8 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import techcourse.myblog.web.dto.UserDto;
+import techcourse.myblog.dto.UserDto;
+import techcourse.myblog.service.UserService;
 
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
@@ -18,6 +19,20 @@ public class UserController {
         this.userService = userService;
     }
 
+    @PostMapping("/users")
+    public String createUser(@Valid UserDto userDto, BindingResult bindingResult, Model model, HttpSession session) {
+        try {
+            if (isLoggedIn(session)) {
+                return "redirect:/";
+            }
+            userService.save(userDto, bindingResult);
+            return "redirect:/login";
+        } catch (IllegalArgumentException e) {
+            model.addAttribute("errorMessage", e.getMessage());
+            return "signup";
+        }
+    }
+
     @GetMapping("/login")
     public String loginView(HttpSession session) {
         if (isLoggedIn(session)) {
@@ -27,25 +42,11 @@ public class UserController {
     }
 
     @GetMapping("/signup")
-    public String sighupView(HttpSession session) {
+    public String signUpView(HttpSession session) {
         if (isLoggedIn(session)) {
             return "redirect:/";
         }
         return "signup";
-    }
-
-    @PostMapping("/users")
-    public String registerUserData(@Valid UserDto userDto, BindingResult bindingResult, Model model, HttpSession session) {
-        try {
-            if (isLoggedIn(session)) {
-                return "redirect:/";
-            }
-            userService.save(userDto, bindingResult);
-            return "redirect:/login";
-        }catch (IllegalArgumentException e) {
-            model.addAttribute("errorMessage", e.getMessage());
-            return "signup";
-        }
     }
 
     private boolean isLoggedIn(HttpSession session) {
